[Refactor] 디자인 패턴 수정
