+++ src/components/Curtain/Curtain.styled.ts
