+++ src/components/Curtain/index.tsx
@@ -1,13 +1,1 @@
-import { MotionValue } from 'framer-motion';
-
-import { CurtainContainer } from './style';
-
-interface CurtainValue {
-  progress: MotionValue<number>;
-}
-
-const Curtain = ({ progress }: CurtainValue) => {
-  return <CurtainContainer style={{ right: progress }} />;
-};
-
-export default Curtain;
+export { default as Curtain } from './Curtain';
