+++ src/components/Modal/ModalFrame.styled.ts
