+++ src/components/Modal/index.tsx
@@ -1,0 +1,1 @@
+export { default as ModalRoot } from './ModalRoot';
