+++ src/components/Sections/index.tsx
@@ -1,0 +1,3 @@
+export { default as EmptySection } from './EmptySection';
+export { default as InputSection } from './InputSection';
+export { default as SelectSection } from './SelectSection';
