+++ src/pages/MainPage.styled.ts
